+++ spec/component-state.decorator.spec.ts
@@ -1,24 +1,41 @@
 import { ComponentState } from '../src/decorators/component-state.decorator';
 import { ServiceLocator } from '../src/helpers/service-locator';
 import { IS_PROD } from '../src/ng-state.module';
+import { Message, Dispatcher } from '../src/services/dispatcher';
 
+const actionId = 'actionId';
 class TestStateActions {
+    aId = actionId;
     createStore(statePath: string[], stateIndex: number | null) {
         return ['newStatePath'];
     }
-};
+}
 
 class TargetComponent {
     statePath: string[];
     stateIndex: number | null;
     actions: any;
-    ngOnInit() { };
+    cd: { markForCheck: () => void };
+
+    ngOnInit() { }
 }
+
+const dispatcher = new Dispatcher();
 
 ServiceLocator.injector = <any>{
     get: (key) => {
         if (key === IS_PROD) {
             return false;
+        }
+
+        if (key.name === 'Dispatcher') {
+            return dispatcher;
+        }
+
+        if (key.name === 'ChangeDetectorRef') {
+            return {
+                markForCheck: () => { }
+            };
         }
     }
 };
@@ -44,5 +61,21 @@ describe('ComponentState decorator', () => {
         target.ngOnInit();
         expect(target.statePath[0]).toBe('newStatePath');
         expect(target.actions instanceof TestStateActions).toBeTruthy();
+    });
+
+    it('should call markForCheck after state change', () => {
+        beforeEach(TestStateActions);
+        target.ngOnInit();
+        jest.spyOn(target.cd, 'markForCheck');
+
+        dispatcher.publish(new Message(actionId, ''));
+
+        expect(target.cd.markForCheck).toHaveBeenCalled();
+    });
+
+    it('should ensure change detector ref is injected', () => {
+        beforeEach(TestStateActions);
+        target.ngOnInit();
+        expect(target.cd.markForCheck).toBeDefined();
     });
 });