+++ spec/localstorage.mock.ts
@@ -1,5 +1,5 @@
 class LocalStorageMock {
-    store = {};
+    store = {} as any;
 
     clear() {
         this.store = {};
@@ -18,4 +18,4 @@ class LocalStorageMock {
     }
 }
 
-Object.defineProperty(window, 'localStorage', { value: new LocalStorageMock() });
+Object.defineProperty(window, 'customStorage', { value: new LocalStorageMock() });