+++ spec/storage.spec.ts
@@ -1,23 +1,26 @@
-import { StateHistory } from './../src/state/history';
+import { StateKeeper } from '../src/state/history';
 import { Store } from './../src/store/store';
 import { NgStateTestBed } from '../src/ng-state.test-bed';
 import { timer } from 'rxjs';
 import { tap, map } from 'rxjs/operators';
-import { PersistStateStorage } from '../src/store/plugins/persist-state.plugin';
+import { PersistStateStorage, PersistStateManager } from '../src/store/plugins/persist-state.plugin';
 
 jest.useFakeTimers();
 
 describe('Storage', () => {
     let store: Store<any>;
+    let keyValueStorage;
 
     beforeEach(() => {
         const initialState = { layout: { test: 'test' } };
         store = NgStateTestBed.createStore(initialState);
+        keyValueStorage = window['customStorage'];
+        PersistStateManager.configureStorage(keyValueStorage, () => Object.keys((keyValueStorage as unknown as LocalStorageMock).store));
     });
 
     it('should save state', () => {
         store.select(['layout']).storage.save({ key: 'testKey' });
-        expect(<any>localStorage.getItem('state::testKey')).toBe('{"test":"test"}');
+        expect(<any>keyValueStorage.getItem('state::testKey')).toBe('{"test":"test"}');
     });
 
     it('should load state', () => {
@@ -25,34 +28,34 @@ describe('Storage', () => {
 
         layoutStore.storage.save();
         layoutStore.update(state => state.set('test', 'test-updated'));
-        expect(StateHistory.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test-updated');
+        expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test-updated');
 
         layoutStore.storage.load();
-        expect(StateHistory.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test');
+        expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test');
     });
 
     it('should clear state', () => {
-        localStorage.setItem('should-stay-item', 'a');
+        keyValueStorage.setItem('should-stay-item', 'a');
         store.select(['layout']).storage.save({ key: 'testKey' });
         store.select(['layout']).storage.clear();
 
-        expect(<any>localStorage.getItem('should-stay-item')).toEqual('a');
+        expect(<any>keyValueStorage.getItem('should-stay-item')).toEqual('a');
     });
 
     it('should remove item', () => {
         store.select(['layout']).storage.save({ key: 'remove-item' });
         store.select(['layout']).storage.remove({ key: 'remove-item' });
 
-        expect(<any>localStorage.getItem('remove-item')).toBeNull();
+        expect(<any>keyValueStorage.getItem('remove-item')).toBeNull();
     });
 
     describe('when delayed', () => {
         const delay = 2000;
         let storage = {
-            clear: () => timer(delay).pipe(tap(_ => localStorage.clear())),
-            getItem: (key: string) => timer(delay).pipe(map(_ => localStorage.getItem(key))),
-            removeItem: (key: string) => timer(delay).pipe(tap(_ => localStorage.removeItem(key))),
-            setItem: (key: string, value: any) => timer(delay).pipe(tap(_ => localStorage.setItem(key, value))),
+            clear: () => timer(delay).pipe(tap(_ => keyValueStorage.clear())),
+            getItem: (key: string) => timer(delay).pipe(map(_ => keyValueStorage.getItem(key))),
+            removeItem: (key: string) => timer(delay).pipe(tap(_ => keyValueStorage.removeItem(key))),
+            setItem: (key: string, value: any) => timer(delay).pipe(tap(_ => keyValueStorage.setItem(key, value))),
         } as PersistStateStorage;
 
         it('should notify observer after state is saved', (done) => {
@@ -61,11 +64,11 @@ describe('Storage', () => {
                     key: 'testKey',
                     storageConfig: {
                         storage: storage,
-                        getKeys: () => Object.keys((localStorage as unknown as LocalStorageMock).store)
+                        getKeys: () => Object.keys((keyValueStorage as unknown as LocalStorageMock).store)
                     }
                 }).subscribe(data => {
                     expect(data.key).toEqual('state::testKey');
-                    expect(data.data).toMatchObject({'test': 'test'});
+                    expect(data.data).toMatchObject({ 'test': 'test' });
                     done();
                 });
 
@@ -75,18 +78,18 @@ describe('Storage', () => {
         it('should notify observer after state is loaded', (done) => {
             const layoutStore = store.select(['layout']);
 
-            layoutStore.storage.save({key: 'testKey'});
+            layoutStore.storage.save({ key: 'testKey' });
 
             layoutStore
                 .storage.load({
                     key: 'testKey',
                     storageConfig: {
                         storage: storage,
-                        getKeys: () => Object.keys((localStorage as unknown as LocalStorageMock).store)
+                        getKeys: () => Object.keys((keyValueStorage as unknown as LocalStorageMock).store)
                     }
                 }).subscribe(data => {
                     expect(data.key).toEqual('state::testKey');
-                    expect(JSON.parse(data.data)).toMatchObject({'test': 'test'});
+                    expect(JSON.parse(data.data)).toMatchObject({ 'test': 'test' });
                     done();
                 });
 
@@ -96,14 +99,14 @@ describe('Storage', () => {
         it('should notify observer after item is removed from storage', (done) => {
             const layoutStore = store.select(['layout']);
 
-            layoutStore.storage.save({key: 'testKey'});
+            layoutStore.storage.save({ key: 'testKey' });
 
             layoutStore
                 .storage.remove({
                     key: 'testKey',
                     storageConfig: {
                         storage: storage,
-                        getKeys: () => Object.keys((localStorage as unknown as LocalStorageMock).store)
+                        getKeys: () => Object.keys((keyValueStorage as unknown as LocalStorageMock).store)
                     }
                 }).subscribe(key => {
                     expect(key).toEqual('state::testKey');
@@ -116,14 +119,14 @@ describe('Storage', () => {
         it('should notify observer after storage is cleared', (done) => {
             const layoutStore = store.select(['layout']);
 
-            layoutStore.storage.save({key: 'testKey'});
-            layoutStore.storage.save({key: 'testKey2'});
+            layoutStore.storage.save({ key: 'testKey' });
+            layoutStore.storage.save({ key: 'testKey2' });
 
             layoutStore
                 .storage.clear({
                     storageConfig: {
                         storage: storage,
-                        getKeys: () => Object.keys((localStorage as unknown as LocalStorageMock).store)
+                        getKeys: () => Object.keys((keyValueStorage as unknown as LocalStorageMock).store)
                     }
                 }).subscribe(keys => {
                     expect(keys.length).toBe(2);