+++ spec/store.spec.ts
@@ -1,4 +1,4 @@
-import { StateHistory } from './../src/state/history';
+import { StateKeeper } from '../src/state/history';
 import { Store } from './../src/store/store';
 import { stateFactory } from '../src/ng-state.module';
 import { BehaviorSubject } from 'rxjs';
@@ -31,14 +31,14 @@ describe('Store tests', () => {
         it('should initialize state with initial value', () => {
             store.initialize([], { test: 'test' });
 
-            expect(StateHistory.CURRENT_STATE.get('test')).toEqual('test');
-            expect(StateHistory.CURRENT_STATE.get('__initialized')).toEqual(true);
+            expect(StateKeeper.CURRENT_STATE.get('test')).toEqual('test');
+            expect(StateKeeper.CURRENT_STATE.get('__initialized')).toEqual(true);
         });
 
         it('should update state', () => {
             store.select(['layout']).update(state => state.set('loading', true));
 
-            expect(StateHistory.CURRENT_STATE.getIn(['layout', 'loading'])).toEqual(true);
+            expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'loading'])).toEqual(true);
         });
 
         it('should select state', (done) => {
@@ -53,29 +53,29 @@ describe('Store tests', () => {
         it('should clear state', () => {
             store.initialize(['router'], { url: 'home' }, false);
             store.select(['layout']).update(state => state.set('loading', true));
-            expect(StateHistory.CURRENT_STATE.getIn(['layout', 'loading'])).toEqual(true);
+            expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'loading'])).toEqual(true);
 
-            store.clear();
-            expect(StateHistory.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test');
-            expect(StateHistory.CURRENT_STATE.getIn(['layout', 'loading'])).not.toBeDefined();
-            expect(StateHistory.CURRENT_STATE.getIn(['router', 'url'])).toBe('');
+            store.reset();
+            expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test');
+            expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'loading'])).not.toBeDefined();
+            expect(StateKeeper.CURRENT_STATE.getIn(['router', 'url'])).toBe('');
         });
 
         it('should reset state', () => {
             store.select(['layout']).update(state => state.set('test', 'test2'));
-            expect(StateHistory.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test2');
+            expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test2');
 
             store.select(['layout']).reset();
-            expect(StateHistory.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test');
+            expect(StateKeeper.CURRENT_STATE.getIn(['layout', 'test'])).toEqual('test');
         });
 
         it('should reset state when stroe has initial state', () => {
             const intilizedStore = store.initialize(['actionStore'], { test: { url: 'home' } });
             intilizedStore.select(['test']).update(state => state.set('url', 'home-updated'));
-            expect(StateHistory.CURRENT_STATE.getIn(['actionStore', 'test', 'url'])).toEqual('home-updated');
+            expect(StateKeeper.CURRENT_STATE.getIn(['actionStore', 'test', 'url'])).toEqual('home-updated');
 
             intilizedStore.select(['test']).reset();
-            expect(StateHistory.CURRENT_STATE.getIn(['actionStore', 'test', 'url'])).toEqual('home');
+            expect(StateKeeper.CURRENT_STATE.getIn(['actionStore', 'test', 'url'])).toEqual('home');
         });
 
         it('should rootPath and initialState vallues to store after initialization', () => {