+++ src/store/reset.ts
@@ -3,6 +3,7 @@ import { StateHistory } from '../state/history';
 import { Map, fromJS } from 'immutable';
 import { tap, take } from 'rxjs/operators';
 import { Helpers } from '../helpers/helpers';
+import { RouterState } from '../state/router-state';
 
 export class Reset {
     constructor() {
@@ -11,15 +12,26 @@ export class Reset {
         const restoreState = function (store: Store<any>) {
             store
                 .update((state: Map<any, any>) => {
+                    let path = store.statePath.filter(item => !store.rootPath.includes(item));
+                    const isRootPath = Array.isArray(path) && path.length === 0;
+
+                    let router = '';
+                    if (isRootPath) {
+                        router = state.get('router');
+                    }
+
                     state.clear();
 
                     let initialState: Map<any, any> = !!store.initialState
                         ? store.initialState
                         : fromJS(StateHistory.initialState);
 
-                    let path = store.statePath.filter(item => !store.rootPath.includes(item));
-
                     state.merge(initialState.getIn(path));
+
+                    if (isRootPath) {
+                        state.set('router', router);
+                        state.setIn(['router', 'url'], RouterState.startingRoute);
+                    }
                 });
         };