+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 5.0.0
+- Added integration with Redux DevTools - [documentation](https://vytautaspranskunas.gitbook.io/react-state-rxjs/debugging/redux-devtools)
+- Improved async support in actions - [documentation](https://vytautaspranskunas.gitbook.io/react-state-rxjs/core-concepts/actions/async)
+- Improved debugging experience
+
+##### BREAKING CHANGES
+- Removed clear operator. Instead use immutable operator ```store.update(state => state.clear())```. Previouse functionality of clear operator was moved to ```reset``` to avoid confusion
+- Time travel was removed because of integration with Redux DevTools. You do not need to include <StateHistoryComponent /> to your main.tsx file.
+
+
 ### 4.4.0
 - persist state plugin return observable on `save`, `load`, `removeItem`, `clear` actions.
 - `onChange` and `shouldUpdateState` hooks are added to form manager plugin